A: Abstracted note list and its item ViewMvc
